#18 refactor : 팀 컨벤션 따라 코드 구조 수정
