--- Aos/app/src/main/java/com/avengers/nibobnebob/presentation/ui/main/mypage/MyPageUiState.kt
@@ -1,8 +1,0 @@
-package com.avengers.nibobnebob.presentation.ui.main.mypage
-
-data class MyPageUiState(
-    val image: String = "",
-    val nickName: String = "",
-    val age: String = "",
-    val area: String = ""
-)
