--- Aos/app/src/main/java/com/avengers/nibobnebob/presentation/ui/main/mypage/edit_profile/EditProfileUiEvent.kt
@@ -1,5 +1,0 @@
-package com.avengers.nibobnebob.presentation.ui.main.mypage.edit_profile
-
-sealed interface EditProfileUiEvent {
-
-}
