+++ Aos/app/src/main/java/com/avengers/nibobnebob/presentation/ui/main/mypage/my_list/MyRestaurantListFragment.kt
@@ -10,8 +10,8 @@ import com.avengers.nibobnebob.R
 import com.avengers.nibobnebob.databinding.FragmentMyRestaurantListBinding
 import com.avengers.nibobnebob.presentation.base.BaseFragment
 import com.avengers.nibobnebob.presentation.ui.main.MainViewModel
-import com.avengers.nibobnebob.presentation.ui.main.mypage.share.MyPageSharedViewModel
 import com.avengers.nibobnebob.presentation.ui.main.mypage.share.MyPageSharedUiEvent
+import com.avengers.nibobnebob.presentation.ui.main.mypage.share.MyPageSharedViewModel
 
 class MyRestaurantListFragment :
     BaseFragment<FragmentMyRestaurantListBinding>(R.layout.fragment_my_restaurant_list) {
